Add TODO about default behavior in call-to-return
When we are not going to analyze 'callee', we do not want to lose all the marks on the instance, arguments, etc. The removal of marks on 'instance'/args (in call-to-return) is only makes sense (and is necessary) when we are truly going to analyze the 'callee', where the removed facts have a possibility to return to returnSite via (summary and) exit-to-return flow function.